requirements.txt
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Admin==1.5.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.3.0
pycparser==2.19
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
WTForms==2.2.1
